backend/requirements.txt
pydantic-ai-slim[openai,anthropic]>=0.2.0
duckdb>=0.9.0
plotly>=5.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
fastapi>=0.100.0
uvicorn>=0.30.0
